rss/pyproject.toml
[project]
name = "rss"
version = "0.1.0"
description = "RSS feed aggregator"
authors = [{ name = "0xsj", email = "sjtommylee@gmail.com" }]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "python-dotenv>=1.0.1",
    "protobuf>=5.27.1",
    "grpcio>=1.64.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = { cmd = "uvicorn src.server:app --host 0.0.0.0 --port 8000 --reload", help = "start server" }
